refactor: moved the core logic separate folder
This allows us to keep a sane separation between the core OIDC logic
and the orchestration services
