+++ projects/angular-simple-oidc/core/lib/errors.ts
