+++ projects/angular-simple-oidc/core/lib/token-crypto.service.ts
