+++ projects/angular-simple-oidc/core/lib/token-helper.service.ts
