+++ projects/angular-simple-oidc/core/lib/token-validation.service.ts
