+++ projects/angular-simple-oidc/core/lib/validate-object-required-props.spec.ts
