+++ projects/angular-simple-oidc/core/package.json
@@ -1,0 +1,10 @@
+{
+    "ngPackage": {
+        "lib": {
+            "entryFile": "public-api.ts",
+            "umdModuleIds": {
+                "jsrsasign": "jsrsasign"
+            }
+        }
+    }
+}
