+++ projects/angular-simple-oidc/src/angular-simple-oidc.module.ts
@@ -1,19 +1,19 @@
 import { NgModule, ModuleWithProviders } from '@angular/core';
-import { AUTH_CONFIG, WINDOW_REF, LOCAL_STORAGE_REF } from './constants';
-import { AuthConfig } from './config/models';
-import { AuthConfigService } from './config/auth-config.service';
-import { SIMPLE_OIDC_APP_INITIALIZER } from './auth.bootstrap';
-import { OidcDiscoveryDocClient } from './discovery-document/oidc-discovery-doc-client.service';
-import { OidcCodeFlowClient } from './oidc-code-flow-client.service';
-import { AuthGuard } from './guards/auth.guard';
-import { AuthService } from './auth.service';
-import { AngularSimpleOidcCoreModule } from './core/angular-simple-oidc-core.module';
-import { TokenStorageService } from './token-storage.service';
-import { TokenEndpointClientService } from './token-endpoint-client.service';
+import { AUTH_CONFIG, WINDOW_REF, LOCAL_STORAGE_REF } from './lib/constants';
+import { AuthConfig } from './lib/config/models';
+import { AuthConfigService } from './lib/config/auth-config.service';
+import { SIMPLE_OIDC_APP_INITIALIZER } from './lib/auth.bootstrap';
+import { OidcDiscoveryDocClient } from './lib/discovery-document/oidc-discovery-doc-client.service';
+import { OidcCodeFlowClient } from './lib/oidc-code-flow-client.service';
+import { AuthGuard } from './lib/guards/auth.guard';
+import { AuthService } from './lib/auth.service';
+import { AngularSimpleOidcCoreModule } from 'angular-simple-oidc/core';
+import { TokenStorageService } from './lib/token-storage.service';
+import { TokenEndpointClientService } from './lib/token-endpoint-client.service';
 import { HttpClientModule } from '@angular/common/http';
-import { RefreshTokenClient } from './refresh-token-client.service';
-import { TokenEventsModule } from './token-events.module';
-import { EndSessionClientService } from './end-session-client.service';
+import { RefreshTokenClient } from './lib/refresh-token-client.service';
+import { TokenEventsModule } from './lib/token-events.module';
+import { EndSessionClientService } from './lib/end-session-client.service';
 
 @NgModule({
   imports: [
