--- projects/angular-simple-oidc/src/lib/core/angular-simple-oidc-core.module.ts
@@ -1,20 +1,0 @@
-import { NgModule } from '@angular/core';
-import { TokenCryptoService } from './token-crypto.service';
-import { TokenHelperService } from './token-helper.service';
-import { TokenValidationService } from './token-validation.service';
-import { TokenUrlService } from './token-url.service';
-import { RefreshTokenValidationService } from './refresh-token/refresh-token-validation.service';
-
-@NgModule({
-  imports: [
-  ],
-  providers: [
-    TokenCryptoService,
-    TokenUrlService,
-    TokenHelperService,
-    TokenValidationService,
-    RefreshTokenValidationService,
-  ],
-  declarations: [],
-})
-export class AngularSimpleOidcCoreModule { }
