+++ projects/angular-simple-oidc/src/lib/token-storage.service.ts
@@ -1,8 +1,10 @@
 import { Injectable, Inject } from '@angular/core';
 import {
-    TokenStorageKeys, LocalState,
-    TokenRequestResult, DecodedIdentityToken
-} from './core/models';
+    TokenStorageKeys,
+    LocalState,
+    TokenRequestResult,
+    DecodedIdentityToken
+} from 'angular-simple-oidc/core';
 import { of, BehaviorSubject } from 'rxjs';
 import { LOCAL_STORAGE_REF } from './constants';
